backend/requirements.txt
django
django-cors-headers
numpy
pyserial
websocket-client
channels
daphne